Moved all libs and exe into src, added lib for transcoder.
