+++ src/encoder/basisu_frontend.cpp
