+++ src/encoder/basisu_ssim.cpp
