+++ src/encoder/lodepng.cpp
