+++ src/encoder/lodepng.h
